Inverse Gauss work!

+++ Estimation.h
@@ -2,9 +2,9 @@
 #include "Condition.h"
 
 //find octahedral norme of vector
-double	octah_vect_norm(double *x, int size)
+float	octah_vect_norm(float *x, int size)
 {
-	double	norm = 0;
+	float	norm = 0;
 
 	for (int i = 0; i < size; i++)
 		norm += fabs(x[i]);
@@ -12,9 +12,9 @@ double	octah_vect_norm(double *x, int size)
 }
 
 //find cube norme of vector
-double	cube_vect_norm(double *x, int size)
+float	cube_vect_norm(float *x, int size)
 {
-	double	norm = 0;
+	float	norm = 0;
 
 	for (int i = 0; i < size; i++)
 		if (norm < x[i])
@@ -23,9 +23,9 @@ double	cube_vect_norm(double *x, int size)
 }
 
 //find sferical norm of vector
-double	sfer_vect_norm(double *x, int DIM)
+float	sfer_vect_norm(float *x, int DIM)
 {
-	double	norm = 0;
+	float	norm = 0;
 
 	for (int i = 0; i < DIM; i++)
 		norm += x[i] * x[i];
@@ -33,9 +33,9 @@ double	sfer_vect_norm(double *x, int DIM)
 }
 
 //find absolutely difference of two vectors
-double	*abs_diff_vector(double *a, double *b, int DIM)
+float	*abs_diff_vector(float *a, float *b, int DIM)
 {
-	double	*c = new double[DIM];
+	float	*c = new float[DIM];
 
 	for (int i = 0; i < DIM; i++)
 		c[i] = fabs(a[i] - b[i]);
@@ -43,9 +43,9 @@ double	*abs_diff_vector(double *a, double *b, int DIM)
 }
 
 //find difference of two vectors
-double	*diff_vector(double *a, double *b, int DIM)
+float	*diff_vector(float *a, float *b, int DIM)
 {
-	double	*c = new double[DIM];
+	float	*c = new float[DIM];
 
 	for (int i = 0; i < DIM; i++)
 		c[i] = a[i] - b[i];
@@ -55,14 +55,14 @@ double	*diff_vector(double *a, double *b, int DIM)
 //find conditional number's estimate
 
 //for cube norme
-double	cube_estimate_number(double *b, const Matrix& R, const Matrix& T)
+float	cube_estimate_number(float *b, const Matrix& R, const Matrix& T)
 {
-	double	pert = 0.1;
-	double	*b_, *tmp = new double[T.size];
-	double	*x, *x_;
-	double	*dx;
-	double	my_dx = 0;
-	double	*db;
+	float	pert = 0.1;
+	float	*b_, *tmp = new float[T.size];
+	float	*x, *x_;
+	float	*dx;
+	float	my_dx = 0;
+	float	*db;
 
 	b_ = multi_vect(b, T);
 	x = find_x(R, b_);
@@ -85,15 +85,15 @@ double	cube_estimate_number(double *b, const Matrix& R, const Matrix& T)
 }
 
 //for octahedral norme
-double	octah_estimate_number(double *b, const Matrix& R, const Matrix& T)
+float	octah_estimate_number(float *b, const Matrix& R, const Matrix& T)
 {
-	double	pert = 0.1;
+	float	pert = 0.1;
 	int DIM = T.size;
-	double	*b_, *tmp = new double[DIM];
-	double	*x, *x_;
-	double	*dx;
-	double	my_dx = 0;
-	double	*db;
+	float	*b_, *tmp = new float[DIM];
+	float	*x, *x_;
+	float	*dx;
+	float	my_dx = 0;
+	float	*db;
 
 	b_ = multi_vect(b, T);
 	x = find_x(R, b_);
@@ -115,15 +115,15 @@ double	octah_estimate_number(double *b, const Matrix& R, const Matrix& T)
 }
 
 //for sferical norm
-double	sfer_estimate_number(double *b, const Matrix& R, const Matrix& T)
+float	sfer_estimate_number(float *b, const Matrix& R, const Matrix& T)
 {
-	double	pert = 0.1;
+	float	pert = 0.1;
 	int DIM = T.size;
-	double	*b_, *tmp = new double[DIM];
-	double	*x, *x_;
-	double	*dx;
-	double	my_dx = 0;
-	double	*db;
+	float	*b_, *tmp = new float[DIM];
+	float	*x, *x_;
+	float	*dx;
+	float	my_dx = 0;
+	float	*db;
 
 	b_ = multi_vect(b, T);
 	x = find_x(R, b_);
